Moved personal settings from about dialog to profile dialog
Profile dialog can now be opened without logging in/without multi-user mode

Fixed issue where about dialog could be accessed from anywhere

Misc internationalization improvements
