+++ src/app/dialogs/about-dialog/about-dialog.component.html
@@ -35,36 +35,6 @@
         <p>
             <ng-container i18n="About bug prefix">Found a bug or have a suggestion?</ng-container>&nbsp;<a [href]="issuesLink" target="_blank"><ng-container i18n="About bug click here">Click here</ng-container></a>&nbsp;<ng-container i18n="About bug suffix">to create an issue!</ng-container>
         </p>
-        <mat-divider></mat-divider>
-        <div style="margin-top: 10px;">
-            <h5>Personal settings:</h5>
-            <mat-form-field>
-                <mat-label i18n="Sidepanel mode">Sidepanel mode</mat-label>
-                <mat-select [(ngModel)]="sidepanel_mode" (selectionChange)="sidePanelModeChanged($event.value)">
-                    <mat-option value="over">
-                        Over
-                    </mat-option>
-                    <mat-option value="side">
-                        Side
-                    </mat-option>
-                </mat-select>
-            </mat-form-field>
-            <br/>
-            <mat-form-field>
-                <mat-label i18n="File card size">File card size</mat-label>
-                <mat-select [(ngModel)]="card_size" (selectionChange)="cardSizeOptionChanged($event.value)">
-                    <mat-option value="large">
-                        Large
-                    </mat-option>
-                    <mat-option value="medium">
-                        Medium
-                    </mat-option>
-                    <mat-option value="small">
-                        Small
-                    </mat-option>
-                </mat-select>
-            </mat-form-field>
-        </div>
     </div>
 </mat-dialog-content>
