+++ src/app/dialogs/about-dialog/about-dialog.component.ts
@@ -16,8 +16,6 @@ export class AboutDialogComponent implements OnInit {
   checking_for_updates = true;
 
   current_version_tag = CURRENT_VERSION;
-  sidepanel_mode = this.postsService.sidepanel_mode;
-  card_size = this.postsService.card_size;
 
   constructor(public postsService: PostsService) { }
 
@@ -31,15 +29,4 @@ export class AboutDialogComponent implements OnInit {
       this.latestGithubRelease = res;
     });
   }
-
-  sidePanelModeChanged(new_mode) {
-    localStorage.setItem('sidepanel_mode', new_mode);
-    this.postsService.sidepanel_mode = new_mode;
-  }
-
-  cardSizeOptionChanged(new_size) {
-    localStorage.setItem('card_size', new_size);
-    this.postsService.card_size = new_size;
-  }
-
 }
